Create structure to dynamically build aggregate queries with InfluxDB

+++ backend/src/main/kotlin/iot/poc/backend/dto/SensorData.kt
@@ -1,10 +1,11 @@
 package iot.poc.backend.dto
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class SensorData(
     val sensorType: String,
     val timestamp: Instant,
     val tags: Map<String, String>,
-    val values: Map<String, Long>
+    val values: Map<String, BigDecimal>
 )
